Extract utils from main source file
